+++ src/main/web-auth/WebAuth.ts
@@ -545,7 +545,7 @@ export class WebAuth {
    * @param options 
    * @returns 
    */
-  getInviteUserDetails(options: { invite_id: string }) {
+  getInviteUserDetails(options: { invite_id: string, callLatestAPI?: boolean }) {
     return UserService.getInviteUserDetails(options);
   };